Money-Map-main/requirements.txt
pymongo
streamlit
bcrypt
pandas
plotly
Pillow

